// src/homeconnect-types.ts
export type Value = string | number | boolean;
// ...
export interface EventData {
    key: string;
    name?: string;
    uri?: string;
    timestamp: number | null;
    level: EventLevel;
    handling: EventHandling;
    value: Value | null;
    displayvalue?: string;
    unit?: string;
    haId?: string;
}
type EventLevel = 'critical' | 'alert' | 'warning' | 'hint' | 'info';
type EventHandling = 'none' | 'acknowledge' | 'decision';
// ...
export interface StatusesWrapper {
    data: {
        status: Status[];
    };
}
export interface StatusWrapper {
    data: Status;
}
export type Status = StatusString | StatusNumber | StatusBoolean;
export interface StatusString extends StatusCommon {
    value: string;
    constraints?: ConstraintsString;
}
export interface StatusNumber extends StatusCommon {
    value: number;
    constraints?: ConstraintsNumber;
}
export interface StatusBoolean extends StatusCommon {
    value: boolean;
    constraints?: ConstraintsBoolean;
}